pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "genai-research"
version = "0.1.0"
description = "Research infrastructure: TinyLlama inference and prompt strategies"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.0.0",
  "transformers>=4.40.0",
  "accelerate>=0.30.0",
  "safetensors>=0.4.0",
  "pyyaml>=6.0.0",
  "datasets>=2.19.0",
  "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "ruff>=0.4.0"]

[project.scripts]
genai-llm = "genai_research.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
